% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceR-package.R
\docType{package}
\name{elevation_raster}
\alias{elevation_raster}
\title{elevation_raster: A package that is helpful for creating surface features such as slope and aspect for geospatial analysis using \code{R}.}
\description{
This package provides functions for generating surface features like slope and aspect from digital elevation models. It is designed for geospatial analysis in R.
}
\references{
\url{https://github.com/p/ChinyereRuth/surfaceR/}
}
\author{
Chinyere Ottah, School of Earth, Environment and Society, McMaster University \email{ottahc@mcmaster.ca}
}
