% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceR-package.R
\docType{package}
\name{surfaceR-package}
\alias{surfaceR}
\alias{surfaceR-package}
\title{surfaceR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChinyereRuth/surfaceR}
  \item Report bugs at \url{https://github.com/ChinyereRuth/surfaceR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
